Fix a couple tests and some warning on mips32

+++ tests/core/pipe/test_pipe_read_callback_extra_data.c
@@ -121,7 +121,7 @@ evbuffer_drain(struct evbuffer *buf, size_t len)
         if (len != sizeof(request) - sizeof(REQUEST_1)) {
             errx(
                 1, "trying to drain %ju bytes (expected %ju)",
-                (uintmax_t)len, sizeof(request) - sizeof(REQUEST_1)
+                (uintmax_t)len, (uintmax_t)(sizeof(request) - sizeof(REQUEST_1))
             );
         }
         break;
@@ -143,7 +143,9 @@ evbuffer_pullup(struct evbuffer *buf, ev_ssize_t size)
     }
 
     if (size != -1) {
-        errx(1, "got unexpected size %ld in pullup (expected -1)", size);
+        errx(
+            1, "got unexpected size %jd in pullup (expected -1)", (intmax_t)size
+        );
     }
 
     return &request[evbuffer_pullup_call_count++ ? sizeof(request) - 1 : 0];
